+++ src/models/index.ts
@@ -2241,6 +2241,8 @@ export interface DocumentCreate {
   rejectDuplicates?: boolean;
   /** A JSON representation of the RegionBias object. */
   regionBias?: string;
+  /** Explicitly mark this document as low priority. */
+  lowPriority?: boolean;
 }
 
 export interface OrganizationCreate {
@@ -3601,6 +3603,8 @@ export interface AffindaAPICreateDocumentOptionalParams
   rejectDuplicates?: boolean;
   /** A JSON representation of the RegionBias object. */
   regionBias?: string;
+  /** Explicitly mark this document as low priority. */
+  lowPriority?: boolean;
 }
 
 /** Contains response data for the createDocument operation. */